+++ LessonFor_1/src/ru/itis/Task22C.java
@@ -1,12 +1,17 @@
+//Задание 10(в), стр 22
+
 package ru.itis;
+
 import java.util.Scanner;
+
 public class Task22C {
     public static void main(String[] args){
+
         Scanner in = new Scanner(System.in);
         int n = in.nextInt();
         int k;
         double sum = 0;
-        double x = in.nextInt();
+        double x = in.nextDouble();
         double an = x*x;
         int fact = 1;
         int fact2 = 1;
